// ui/src/generated/bind-type-response.ts
// tslint:disable
/**
 * 
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { DataType } from './data-type';
import { ExpressionData } from './expression-data';
import { ProjectData } from './project-data';
import { Typeclass } from './typeclass';

/**
 * 
 * @export
 * @interface BindTypeResponse
 */
export interface BindTypeResponse {
    /**
     * 
     * @type {ProjectData}
     * @memberof BindTypeResponse
     */
    btProjectData: ProjectData;
    /**
     * 
     * @type {DataType}
     * @memberof BindTypeResponse
     */
    btDataType: DataType;
    /**
     * 
     * @type {string}
     * @memberof BindTypeResponse
     */
    btPrettyType: string;
    /**
     * 
     * @type {ExpressionData}
     * @memberof BindTypeResponse
     */
    btCodegen?: ExpressionData;
    /**
     * 
     * @type {Array<Typeclass>}
     * @memberof BindTypeResponse
     */
    btTypeclasses: Array<Typeclass>;
}