// ui/src/generated/data-type.ts
// tslint:disable
/**
 * 
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { Type } from './type';

/**
 * 
 * @export
 * @interface DataType
 */
export interface DataType {
    /**
     * 
     * @type {string}
     * @memberof DataType
     */
    dtName: string;
    /**
     * 
     * @type {Array<string>}
     * @memberof DataType
     */
    dtVars: Array<string>;
    /**
     * 
     * @type {{ [key: string]: Array<Type>; }}
     * @memberof DataType
     */
    dtConstructors: { [key: string]: Array<Type>; };
}